+++ src/main/scala/factoryMethod/examples/pizzaRestaurant/products/PepperoniPizza.scala
@@ -1,0 +1,5 @@
+package factoryMethod.examples.pizzaRestaurant.products
+
+class PepperoniPizza (override val price:Double) extends Pizza (price){
+
+}